requirements.txt
flask==2.0.1
flask-cors==3.0.10
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0
torchvision==0.10.0
torchaudio==0.9.0
numpy>=1.21.0,<1.25.0
Werkzeug==2.0.1
pytorch-lightning==1.4.0
torchmetrics==0.6.0
mediapipe>=0.8.9.1
face-alignment==1.3.5
opencv-python-headless>=4.6.0
sentencepiece==0.1.99
av>=11.0.0